REFACTOR-002: extract parse issue primitives into ./parse/primitives.ts
Pure structural move, no behavior change. Moves DomainIssueCode,
DomainIssue, ParseResult, DomainValidationError, formatIssue, and the
private issue()/fromZod()/parseWith()/claimId()/checkOrdering() helpers
out of the growing parse.ts monolith into their own module. parse.ts
re-exports the same public surface and imports the moved helpers, so
every consumer keeps importing from ./parse unchanged.

First of a stack splitting parse.ts (REFACTOR-002) into per-concern
modules so a new invariant lands in its own file instead of on the same
883-line wall every parallel PR was editing.

Refs #143
